Code review Phase 2: correctness fixes with regression tests
- Save/restore DAIF across context switches (SCHED-C1)
- Fix timer_percpu_init using wrong IRQ on Pi 5 (DRV-C4)
- Fix VirtIO RX buffer mapping after descriptor recycling (DRV-C5)
- Fix GICv2 SGI target: use bitmask not CPU number (DRV-M1)
- Fix vmm_dump integer overflow on Pi 5 (MM-M6)
- Correct Pi 5 UART_CLOCK from 48 MHz to 50 MHz (DOC-C3)

Regression tests added for DAIF preservation and VMM stats overflow.
All tests pass on QEMU. Boot test 10/10 on Pi 5 hardware.

+++ kernel/mm/vmm.c
@@ -638,7 +638,7 @@ void vmm_dump(void)
     uart_printf("  L1 table at: %p\n", l1_table);
     uart_printf("  Blocks mapped: %u\n", vmm_state.blocks_mapped);
     uart_printf("  Bytes mapped: %lu MB\n",
-                (vmm_state.blocks_mapped * BLOCK_SIZE) / (1024 * 1024));
+                ((uint64_t)vmm_state.blocks_mapped * BLOCK_SIZE) / (1024 * 1024));
 
     /* Dump non-empty L1 entries */
     for (int i = 0; i < ENTRIES_PER_TABLE; i++) {
@@ -832,7 +832,7 @@ void vmm_init(void)
     DEBUG_PRINT("  L2 tables used: %u", vmm_state.l2_tables_used);
     DEBUG_PRINT("  Blocks mapped: %u (%lu MB)",
                 vmm_state.blocks_mapped,
-                (vmm_state.blocks_mapped * BLOCK_SIZE) / (1024 * 1024));
+                ((uint64_t)vmm_state.blocks_mapped * BLOCK_SIZE) / (1024 * 1024));
 
     /*
      * Enable MMU.
